Changed the file structure
