+++ app/src/main/java/com/gb/stopwatch/data/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.gb.stopwatch.data
+
+interface TimestampProvider {
+    fun getMilliseconds(): Long
+}
