+++ app/src/main/java/com/gb/stopwatch/data/TimestampProviderImpl.kt
@@ -1,0 +1,7 @@
+package com.gb.stopwatch.data
+
+class TimestampProviderImpl : TimestampProvider {
+    override fun getMilliseconds(): Long {
+        return System.currentTimeMillis()
+    }
+}
