+++ app/src/main/java/com/gb/stopwatch/view/ElapsedTimeCalculator.kt
@@ -1,0 +1,19 @@
+package com.gb.stopwatch.view
+
+import com.gb.stopwatch.data.TimestampProvider
+import com.gb.stopwatch.stopwatch.StopwatchState
+
+class ElapsedTimeCalculator(
+    private val timestampProvider: TimestampProvider,
+) {
+
+    fun calculate(state: StopwatchState.Running): Long {
+        val currentTimestamp = timestampProvider.getMilliseconds()
+        val timePassedSinceStart = if (currentTimestamp > state.startTime) {
+            currentTimestamp - state.startTime
+        } else {
+            0
+        }
+        return timePassedSinceStart + state.elapsedTime
+    }
+}
